+++ src/modules/default/index.ts
@@ -1,1 +1,6 @@
-export { default as BiliInfo } from './biliInfo'
+export { default as Default_UserInfo } from './userInfo'
+export { default as Default_DailyRewardInfo } from './dailyRewardInfo'
+export { default as Default_DynamicVideos } from './dynamicVideos'
+export { default as Default_FansMetals } from './fansMetals'
+export { default as Default_Cookies } from './cookies'
+export { default as Default_BilibiliLive } from './bilibiliLive'
