+++ src/modules/enhanceExperience/noSleep.ts
@@ -1,10 +1,9 @@
-import { onFrameTypes, runAtMoment } from '../../types/module'
+import { runAtMoment } from '../../types/module'
 import BaseModule from '../BaseModule'
 
 class NoSleep extends BaseModule {
   static runMultiple: boolean = true
   static runAt: runAtMoment = 'window-load'
-  static onFrame: onFrameTypes = 'target'
   static runAfterDefault: boolean = false
 
   config = this.moduleStore.moduleConfig.EnhanceExperience.sleepDetection
