+++ src/types/cookies.d.ts
@@ -1,5 +1,6 @@
 interface IbiliCookies {
   bili_jct: string | null
+  LIVE_BUVID: string | null
 }
 
 export { IbiliCookies }
